+++ src/components/app/app.jsx
@@ -4,9 +4,7 @@ import BurgerIngredients from "../burger-ingredients/burger-ingredients";
 import BurgerConstructor from "../burger-constructor/burger-constructor";
 import { useState, useEffect } from "react";
 import { getDataFetch } from "../../api/api";
-import Modal from "../modal/modal";
-import IngredientDetails from "../ingredient-details/ingredient-details";
-import OrderDetails from "../order-details/order-details";
+
 
 function App() {
   //cостояние для массива из апи
@@ -25,29 +23,6 @@ function App() {
         console.log(err);
       });
   }
-  //состояния отрытия модального окна и ингредиента:
-  const [isOpenIngredient, setIsOpenIngredient] = useState(false);
-  const [tabIngredient, setTabIngredient] = useState(null);
-
-  function handleOpenModalIngredient(item) {
-    setIsOpenIngredient(true);
-    setTabIngredient(item);
-  }
-
-  const handleCloseModalIngredient = () => {
-    setIsOpenIngredient(false);
-    setTabIngredient(null);
-  };
-  //состояния отрытия модального окна для работы попапа заказа:
-  const [isOpen, setIsOpen] = useState(false);
-
-  const handleOpenModal = () => {
-    setIsOpen(true);
-  };
-
-  const handleCloseModal = () => {
-    setIsOpen(false);
-  };
 
   return (
     <div className={styles.app}>
@@ -58,7 +33,6 @@ function App() {
           {burgerIngredients.length && (
             <BurgerIngredients
               ingridients={burgerIngredients}
-              onTab={handleOpenModalIngredient}
             />
           )}
         </section>
@@ -66,21 +40,10 @@ function App() {
           {burgerIngredients.length && (
             <BurgerConstructor
               ingridients={burgerIngredients}
-              onClick={handleOpenModal}
             />
           )}
         </section>
       </main>
-      <Modal
-        isOpen={isOpenIngredient}
-        onClose={handleCloseModalIngredient}
-        title="Детали ингредиента"
-      >
-        <IngredientDetails tabIngredient={tabIngredient} />
-      </Modal>
-      <Modal isOpen={isOpen} onClose={handleCloseModal}>
-        <OrderDetails />
-      </Modal>
     </div>
   );
 }
