+++ src/components/burger-ingredients/burger-ingredients.jsx
@@ -4,12 +4,27 @@ import ingredientsStyles from "./burger-ingredients.module.css";
 import IngridientItem from "../ingridient-item/ingridient-item";
 import { ingredientPropType } from "../../utils/prop-types";
 import PropTypes from "prop-types";
+import Modal from "../modal/modal";
+import IngredientDetails from "../ingredient-details/ingredient-details";
 
-function BurgerIngredients({ ingridients, onTab }) {
+function BurgerIngredients({ ingridients }) {
   const [current, setCurrent] = useState("one");
   const bun = "bun";
   const sauce = "sauce";
   const main = "main";
+  //состояния отрытия модального окна и ингредиента:
+  const [isOpenIngredient, setIsOpenIngredient] = useState(false);
+  const [tabIngredient, setTabIngredient] = useState(null);
+
+  function handleOpenModalIngredient(item) {
+    setIsOpenIngredient(true);
+    setTabIngredient(item);
+  }
+
+  const handleCloseModalIngredient = () => {
+    setIsOpenIngredient(false);
+    setTabIngredient(null);
+  };
   //нашла все булки
   const buns = useMemo(
     () => ingridients.filter((m) => m.type === bun),
@@ -58,7 +73,10 @@ function BurgerIngredients({ ingridients, onTab }) {
           <ul className={`${ingredientsStyles.ingridient__list} pt-5`}>
             {buns.map((ingridients) => (
               <li key={ingridients._id}>
-                <IngridientItem ingridient={ingridients} onTab={onTab} />
+                <IngridientItem
+                  ingridient={ingridients}
+                  onTab={handleOpenModalIngredient}
+                />
               </li>
             ))}
           </ul>
@@ -70,7 +88,10 @@ function BurgerIngredients({ ingridients, onTab }) {
           <ul className={`${ingredientsStyles.ingridient__list} pt-5`}>
             {sauces.map((ingridients) => (
               <li key={ingridients._id}>
-                <IngridientItem ingridient={ingridients} onTab={onTab} />
+                <IngridientItem
+                  ingridient={ingridients}
+                  onTab={handleOpenModalIngredient}
+                />
               </li>
             ))}
           </ul>
@@ -82,12 +103,20 @@ function BurgerIngredients({ ingridients, onTab }) {
           <ul className={`${ingredientsStyles.ingridient__list} pt-5`}>
             {fillings.map((ingridients) => (
               <li key={ingridients._id}>
-                <IngridientItem ingridient={ingridients} onTab={onTab} />
+                <IngridientItem
+                  ingridient={ingridients}
+                  onTab={handleOpenModalIngredient}
+                />
               </li>
             ))}
           </ul>
         </div>
       </div>
+      {isOpenIngredient && (
+       (<Modal onClose={handleCloseModalIngredient} title="Детали ингредиента">
+          <IngredientDetails tabIngredient={tabIngredient} />
+        </Modal>) 
+      )}
     </>
   );
 }
