+++ src/components/modal/modal.jsx
@@ -3,10 +3,11 @@ import ReactDOM from "react-dom";
 import { CloseIcon } from "@ya.praktikum/react-developer-burger-ui-components";
 import ModalOverlay from "../modal-overlay/modal-overlay";
 import modal from "../modal/modal.module.css";
+import PropTypes from 'prop-types';
 
 const modalRoot = document.getElementById("react-modals");
 
-const Modal = ({ children, isOpen, onClose, title }) => {
+const Modal = ({ children, onClose, title }) => {
   useEffect(() => {
     const closeByEscape = (evt) => {
       if (evt.key === "Escape") {
@@ -19,17 +20,13 @@ const Modal = ({ children, isOpen, onClose, title }) => {
     };
   }, [onClose]);
 
-  if (!isOpen) {
-    return null;
-  }
-
   return ReactDOM.createPortal(
     <>
       <div className={`${modal.container} pt-5`}>
         <div className={`${modal.item} pt-5`}>
           <div className={modal.title_close_container}>
             <div className={`${modal.title_container}`}>
-            { title &&
+            { title && 
               <h2 className={`${modal.title} text text_type_main-large`}>{title}</h2>}
             </div>  
             <button onClick={onClose} className={modal.button_close}>
@@ -44,5 +41,11 @@ const Modal = ({ children, isOpen, onClose, title }) => {
     modalRoot
   );
 };
+
+Modal.propTypes = {
+  children: PropTypes.element.isRequired,
+  onClose: PropTypes.func.isRequired,
+  title: PropTypes.string
+}
 
 export default Modal;
